Do not break when a component cannot be rendered. Render template instead
Previous versions would break if the underlying route component could
not be rendered. This wasn't ideal because it would leave you with a
situation where routes _had to_ be able to render in a node process.
This meant any DOM interaction, or worse, any lib that relied on the
DOM would cause a route to fail and would cause the entire plugin to
break.

That's unfortunate since you probably would rather just be told that it
couldn't render and have your index file output as well.

Another use case is private routes. I.e. routes which require login.
In cases like this static rendering is less of an issue because there
is no SEO benefit even if the full page is rendered. We now support
rendering blank templates for redirected URLs which supports this use
case.